move modal functionality to create page
