+++ src/components/TheHeader.js
@@ -1,8 +1,9 @@
 import React from 'react';
 import 'boxicons';
-import { NavLink } from 'react-router-dom';
+import { NavLink } from "react-router-dom";
 
 class TheHeader extends React.Component {
+
     render() {
         return (
               <nav className="pinterest_feed">
@@ -44,8 +45,7 @@ class TheHeader extends React.Component {
                       Home
                     </NavLink>
                     <NavLink to="/today" className="button p-2">Today</NavLink>
-                    <NavLink className="button p-2" to="/AddPin">Create</NavLink> <box-icon name="plus-circle"   color="#928e8e"></box-icon>
-        
+                    <NavLink to="/Create.js" className="hover:bg-black">Create</NavLink>
                     <div className="search-bar bg-gray-200 w-9/12 rounded-full flex gap-x-2 p-3 align-center">
                       <box-icon name="search-alt-2"></box-icon>
                             <p>Search</p>
